Add data validator for ShoppingLists and Shops entities

+++ ShopyList/ShopyList.DataAccess/CQRS/Commands/ShoppingListsCommands/DeleteShoppingListByIdCommand.cs
@@ -10,6 +10,10 @@ namespace ShopyList.DataAccess.CQRS.Commands.ShoppingListsCommands
         public override async Task<ShoppingList> Execute(ShopyListStorageContext context)
         {
             var removeShoppingList = await context.ShoppingLists.SingleOrDefaultAsync(x => x.Id == this.Id);
+            if (removeShoppingList == null)
+            {
+                return null;
+            }
             context.ShoppingLists.Remove(removeShoppingList);
             await context.SaveChangesAsync();
             return removeShoppingList;
